update code to fix swagger warnings
