+++ solution/backend/common/auth.py
@@ -1,6 +1,16 @@
 from django.conf import settings
 from rest_framework import authentication, exceptions
+from drf_spectacular.extensions import OpenApiAuthenticationExtension
 
+class SettingsAuthenticationScheme(OpenApiAuthenticationExtension):
+    target_class = 'common.auth.SettingsAuthentication'
+    name = 'SettingsAuth'
+
+    def get_security_definition(self, auto_schema):
+        return {
+            'type': 'http',
+            'scheme': 'bearer'
+        }
 
 class SettingsUser:
     is_authenticated = False
